refactor: move db management functionality into db module
